feat(remote): trigger export of old remote when linking pollers
