+++ src/CentreonRemote/Domain/Service/ConfigurationWizard/LinkedPollerConfigurationService.php
@@ -8,7 +8,9 @@ require_once _CENTREON_PATH_ . 'www/class/centreonBroker.class.php';
 require_once _CENTREON_PATH_ . 'www/class/centreonConfigCentreonBroker.php';
 require_once _CENTREON_PATH_ . 'www/include/configuration/configGenerate/DB-Func.php';
 
+use Centreon\Domain\Entity\Task;
 use CentreonRemote\Domain\Value\PollerServer;
+use CentreonRemote\Infrastructure\Service\PollerInteractionService;
 use Pimple\Container;
 
 class LinkedPollerConfigurationService
@@ -20,11 +22,15 @@ class LinkedPollerConfigurationService
     /** @var \CentreonDB */
     private $db;
 
+    /** @var PollerInteractionService */
+    private $pollerInteraction;
+
 
     public function __construct(Container $di)
     {
         $this->di = $di;
         $this->db = $di['centreon.db-manager']->getAdapter('configuration_db')->getCentreonDBInstance();
+        $this->pollerInteraction = new PollerInteractionService($di);
     }
 
     /**
@@ -33,20 +39,22 @@ class LinkedPollerConfigurationService
      */
     public function setPollersConfigurationWithServer(array $pollers, PollerServer $remote)
     {
-        $pollerIDs = [];
+        $pollerIDs = array_map(function ($poller) {
+            return $poller->getId();
+        }, $pollers);
+
+        // Before linking the pollers to the new remote, we have to tell the old remote they are no longer linked to it
+        $this->triggerExportForOldRemotes($pollerIDs);
 
         foreach ($pollers as $poller) {
             $pollerID = $poller->getId();
 
             $this->setBrokerOutputOfPoller($pollerID, $remote);
             $this->setPollerRelationToRemote($pollerID, $remote);
-
-            $pollerIDs[] = $pollerID;
         }
 
-        $this->generateConfiguration($pollerIDs);
-        $this->moveConfigurationFiles($pollerIDs);
-        $this->restartPoller($pollerIDs);
+        // Generate configuration for pollers and restart them
+        $this->pollerInteraction->generateAndExport($pollerIDs);
     }
 
     private function setBrokerOutputOfPoller($pollerID, PollerServer $remote)
@@ -70,159 +78,40 @@ class LinkedPollerConfigurationService
         $statement->execute([':id' => $pollerID]);
     }
 
-    private function generateConfiguration(array $pollerIDs)
+    private function triggerExportForOldRemotes($pollerIDs)
     {
-        $centreon = $_SESSION['centreon'];
-        $username = 'unknown';
+        // Get from the database only the pollers that are linked to a remote
+        $idBindString = str_repeat('?,', count($pollerIDs));
+        $idBindString = rtrim($idBindString, ',');
+        $queryPollers = "SELECT id, remote_id FROM nagios_server WHERE id IN({$idBindString}) AND remote_id IS NOT NULL";
+        $remotesStatement = $this->db->query($queryPollers, $pollerIDs);
+        $pollersWithRemote = $remotesStatement->fetchAll(\PDO::FETCH_ASSOC);
+        $alreadyExportedRemotes = [];
 
-        if (isset($centreon->user->name)) {
-            $username = $centreon->user->name;
-        }
+        // For each remote get the currently linked pollers, exclude the ones selected and trigger export
+        foreach ($pollersWithRemote as $poller) {
+            $remoteID = $poller['remote_id'];
 
-        try {
-            // Sync contact groups with ldap
-            $contactGroupObject = new \CentreonContactgroup($this->db);
-            $contactGroupObject->syncWithLdap();
-
-            // Generate configuration
-            $configGenerateObject = new \Generate($this->di);
-
-            foreach ($pollerIDs as $pollerID) {
-                $configGenerateObject->reset();
-                $configGenerateObject->configPollerFromId($pollerID, $username);
-            }
-        } catch(\Exception $e) {
-            throw new \Exception('There was an error generating the configuration for a poller.');
-        }
-    }
-
-    private function moveConfigurationFiles(array $pollerIDs)
-    {
-        $centreon = $_SESSION['centreon'];
-        $centreonBrokerPath = _CENTREON_PATH_ . '/filesGeneration/broker/';
-
-        if (defined('_CENTREON_VARLIB_')) {
-            $centCorePipe = _CENTREON_VARLIB_ . '/centcore.cmd';
-        } else {
-            $centCorePipe = '/var/lib/centreon/centcore.cmd';
-        }
-
-        $tabServer = [];
-        $tabs = $centreon->user->access->getPollerAclConf([
-            'fields'     => ['name', 'id', 'localhost'],
-            'order'      => ['name'],
-            'conditions' => ['ns_activate' => '1'],
-            'keys'       => ['id']
-        ]);
-
-        $brokerObj = new \CentreonConfigCentreonBroker($this->db);
-        $correlationPath = $brokerObj->getCorrelationFile();
-        $localId = getLocalhostId();
-
-        foreach ($tabs as $tab) {
-            if (in_array($tab['id'], $pollerIDs)) {
-                $tabServer[$tab['id']] = [
-                    'id'        => $tab['id'],
-                    'name'      => $tab['name'],
-                    'localhost' => $tab['localhost']
-                ];
-            }
-        }
-
-        foreach ($tabServer as $host) {
-            if ($correlationPath !== false && $localId !== false) {
-                $tmpFilename = $centreonBrokerPath . '/' . $host['id'] . '/correlation_' . $host['id'] . '.xml';
-                $filenameToGenerate = dirname($correlationPath) . '/correlation_' . $host['id'] . '.xml';
-
-                // Delete file
-                if (file_exists($filenameToGenerate)) {
-                    @unlink($filenameToGenerate);
-                }
-                // Copy file
-                if (file_exists($tmpFilename)) {
-                    @copy($tmpFilename, $filenameToGenerate);
-                }
+            if (in_array($remoteID, $alreadyExportedRemotes)) {
+                continue;
             }
 
-            if (in_array($host['id'], $pollerIDs)) {
-                $listBrokerFile = glob($centreonBrokerPath . $host['id'] . "/*.{xml,cfg,sql}", GLOB_BRACE);
+            $alreadyExportedRemotes[] = $remoteID;
 
-                passthru("echo 'SENDCFGFILE:{$host['id']}' >> {$centCorePipe}", $return);
+            // Get all linked pollers of the remote
+            $queryPollersOfRemote = "SELECT id FROM nagios_server WHERE remote_id = {$remoteID}";
+            $linkedStatement = $this->db->query($queryPollersOfRemote);
+            $linkedResults = $linkedStatement->fetchAll(\PDO::FETCH_ASSOC);
+            $linkedPollersOfRemote = array_column($linkedResults, 'id');
 
-                if ($return) {
-                    throw new \Exception(_('Could not write into centcore.cmd. Please check file permissions.'));
-                }
+            // Exclude the selected pollers which are going to another remote
+            $pollerIDsToExport = array_diff($linkedPollersOfRemote, $pollerIDs);
 
-                if (count($listBrokerFile) > 0) {
-                    passthru("echo 'SENDCBCFG:" . $host['id'] . "' >> $centCorePipe", $return);
-
-                    if ($return) {
-                        throw new \Exception(_('Could not write into centcore.cmd. Please check file permissions.'));
-                    }
-                }
-            }
-        }
-    }
-
-    private function restartPoller(array $pollerIDs)
-    {
-        $centreon = $_SESSION['centreon'];
-        $tabServers = [];
-
-        if (defined('_CENTREON_VARLIB_')) {
-            $centCorePipe = _CENTREON_VARLIB_ . '/centcore.cmd';
-        } else {
-            $centCorePipe = '/var/lib/centreon/centcore.cmd';
-        }
-
-        $tabs = $centreon->user->access->getPollerAclConf([
-            'fields'     => ['name', 'id', 'localhost', 'init_script'],
-            'order'      => ['name'],
-            'conditions' => ['ns_activate' => '1'],
-            'keys'       => ['id']
-        ]);
-
-        $broker = new \CentreonBroker($this->db);
-        $broker->reload();
-
-        foreach ($tabs as $tab) {
-            if (in_array($tab['id'], $pollerIDs)) {
-                $tabServers[$tab['id']] = [
-                    'id'          => $tab['id'],
-                    'name'        => $tab['name'],
-                    'localhost'   => $tab['localhost'],
-                    'init_script' => $tab['init_script']
-                ];
-            }
-        }
-
-        foreach ($tabServers as $poller) {
-            if (isset($poller['localhost']) && $poller['localhost'] == 1) {
-                shell_exec("sudo service {$poller['init_script']} restart");
-            } else {
-                if ($fh = @fopen($centCorePipe, 'a+')) {
-                    fwrite($fh, 'RESTART:' . $poller['id'] . "\n");
-                    fclose($fh);
-                } else {
-                    throw new \Exception(_('Could not write into centcore.cmd. Please check file permissions.'));
-                }
-            }
-
-            $restartTimeQuery = "UPDATE `nagios_server` 
-                SET `last_restart` = '" . time() . "' 
-                WHERE `id` = '{$poller['id']}'";
-            $this->db->query($restartTimeQuery);
-        }
-
-        // Find restart actions in modules
-        foreach ($centreon->modules as $key => $value) {
-            $moduleFiles = glob(_CENTREON_PATH_ . 'www/modules/' . $key . '/restart_pollers/*.php');
-
-            if ($value['restart'] && $moduleFiles) {
-                foreach ($moduleFiles as $fileName) {
-                    include $fileName;
-                }
-            }
+            $exportParams = [
+                'server'  => $remoteID,
+                'pollers' => $pollerIDsToExport,
+            ];
+            $this->di['centreon.taskservice']->addTask(Task::TYPE_EXPORT, ['params' => $exportParams]);
         }
     }
 }
